+++ benchmarks/src/main/java/io/grpc/benchmarks/qps/AsyncServer.java
@@ -31,10 +31,30 @@
 
 package io.grpc.benchmarks.qps;
 
-import static io.grpc.benchmarks.qps.Utils.newServer;
+import static io.grpc.testing.integration.Util.loadCert;
+
+import com.google.common.util.concurrent.MoreExecutors;
+import com.google.protobuf.ByteString;
 
 import io.grpc.ServerImpl;
+import io.grpc.Status;
+import io.grpc.stub.StreamObserver;
+import io.grpc.testing.Payload;
+import io.grpc.testing.PayloadType;
+import io.grpc.testing.SimpleRequest;
+import io.grpc.testing.SimpleResponse;
+import io.grpc.testing.TestServiceGrpc;
+import io.grpc.transport.netty.GrpcSslContexts;
+import io.grpc.transport.netty.NettyServerBuilder;
+import io.netty.channel.EventLoopGroup;
+import io.netty.channel.ServerChannel;
+import io.netty.channel.nio.NioEventLoopGroup;
+import io.netty.channel.socket.nio.NioServerSocketChannel;
+import io.netty.handler.ssl.SslContext;
+import io.netty.handler.ssl.SslProvider;
 
+import java.io.File;
+import java.io.IOException;
 import java.util.concurrent.TimeUnit;
 
 /**
@@ -78,5 +98,126 @@ public class AsyncServer {
         }
       }
     });
+  }
+
+  static ServerImpl newServer(ServerConfiguration config) throws IOException {
+    SslContext sslContext = null;
+    if (config.tls) {
+      System.out.println("Using fake CA for TLS certificate.\n"
+          + "Run the Java client with --tls --testca");
+
+      File cert = loadCert("server1.pem");
+      File key = loadCert("server1.key");
+      boolean useJdkSsl = config.transport == ServerConfiguration.Transport.NETTY_NIO;
+      sslContext = GrpcSslContexts.forServer(cert, key)
+          .sslProvider(useJdkSsl ? SslProvider.JDK : SslProvider.OPENSSL)
+          .build();
+    }
+
+    final EventLoopGroup boss;
+    final EventLoopGroup worker;
+    final Class<? extends ServerChannel> channelType;
+    switch (config.transport) {
+      case NETTY_NIO: {
+        boss = new NioEventLoopGroup();
+        worker = new NioEventLoopGroup();
+        channelType = NioServerSocketChannel.class;
+        break;
+      }
+      case NETTY_EPOLL: {
+        try {
+          // These classes are only available on linux.
+          Class<?> groupClass = Class.forName("io.netty.channel.epoll.EpollEventLoopGroup");
+          @SuppressWarnings("unchecked")
+          Class<? extends ServerChannel> channelClass = (Class<? extends ServerChannel>)
+              Class.forName("io.netty.channel.epoll.EpollServerSocketChannel");
+          boss = (EventLoopGroup) groupClass.newInstance();
+          worker = (EventLoopGroup) groupClass.newInstance();
+          channelType = channelClass;
+          break;
+        } catch (Exception e) {
+          throw new RuntimeException(e);
+        }
+      }
+      case NETTY_UNIX_DOMAIN_SOCKET: {
+        try {
+          // These classes are only available on linux.
+          Class<?> groupClass = Class.forName("io.netty.channel.epoll.EpollEventLoopGroup");
+          @SuppressWarnings("unchecked")
+          Class<? extends ServerChannel> channelClass = (Class<? extends ServerChannel>)
+              Class.forName("io.netty.channel.epoll.EpollServerDomainSocketChannel");
+          boss = (EventLoopGroup) groupClass.newInstance();
+          worker = (EventLoopGroup) groupClass.newInstance();
+          channelType = channelClass;
+          break;
+        } catch (Exception e) {
+          throw new RuntimeException(e);
+        }
+      }
+      default: {
+        // Should never get here.
+        throw new IllegalArgumentException("Unsupported transport: " + config.transport);
+      }
+    }
+
+    return NettyServerBuilder
+        .forAddress(config.address)
+        .bossEventLoopGroup(boss)
+        .workerEventLoopGroup(worker)
+        .channelType(channelType)
+        .addService(TestServiceGrpc.bindService(new TestServiceImpl()))
+        .sslContext(sslContext)
+        .executor(config.directExecutor ? MoreExecutors.newDirectExecutorService() : null)
+        .connectionWindowSize(config.connectionWindow)
+        .streamWindowSize(config.streamWindow)
+        .build();
+  }
+
+  private static class TestServiceImpl implements TestServiceGrpc.TestService {
+
+    @Override
+    public void unaryCall(SimpleRequest request, StreamObserver<SimpleResponse> responseObserver) {
+      SimpleResponse response = buildSimpleResponse(request);
+      responseObserver.onValue(response);
+      responseObserver.onCompleted();
+    }
+
+    @Override
+    public StreamObserver<SimpleRequest> streamingCall(
+        final StreamObserver<SimpleResponse> responseObserver) {
+      return new StreamObserver<SimpleRequest>() {
+        @Override
+        public void onValue(SimpleRequest request) {
+          SimpleResponse response = buildSimpleResponse(request);
+          responseObserver.onValue(response);
+        }
+
+        @Override
+        public void onError(Throwable t) {
+          System.out.println("Encountered an error in streamingCall");
+          t.printStackTrace();
+        }
+
+        @Override
+        public void onCompleted() {
+          responseObserver.onCompleted();
+        }
+      };
+    }
+
+    private static SimpleResponse buildSimpleResponse(SimpleRequest request) {
+      if (request.getResponseSize() > 0) {
+        if (!PayloadType.COMPRESSABLE.equals(request.getResponseType())) {
+          throw Status.INTERNAL.augmentDescription("Error creating payload.").asRuntimeException();
+        }
+
+        ByteString body = ByteString.copyFrom(new byte[request.getResponseSize()]);
+        PayloadType type = request.getResponseType();
+
+        Payload payload = Payload.newBuilder().setType(type).setBody(body).build();
+        return SimpleResponse.newBuilder().setPayload(payload).build();
+      }
+      return SimpleResponse.getDefaultInstance();
+    }
   }
 }
